requirements.txt
google-api-core==1.30.0
google-api-python-client==1.12.8
google-auth==1.30.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.7.3
idna==2.9
multidict==4.7.5
websockets==8.1
yarl==1.4.2
python-dateutil==2.8.1
asyncpg==0.23.0
